🔮👨 Checkpoint ./config.json:26900934/14 ./commands/fun/coinflip.js:26900934/696 ./commands/fun/howgay.js:26900934/825 ./commands/help/help.js:26900934/86 ./commands/fun/say.js:26900934/1572

+++ commands/fun/say.js
@@ -7,6 +7,14 @@ module.exports = {
     usage: "say <message>",
 
     async run (client, message, args) {
+      if (!message.guild.me.permissions.has("MANAGE_MESSAGES"))
+      return message.channel.send(
+        "I Don't Have Permission To Use This Command! Manage Messages"
+      );
+    if (!message.member.permissions.has("MANAGE_MESSAGES"))
+      return message.channel.send(
+        `You Don't Have Permission To Use This Command! Manage `
+      );
         let msg;
         let textChannel = message.mentions.channels.first()
         message.delete()
